requirements.txt
python-dotenv
poml==0.0.8
langchain
langchain-google-genai
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
pandas
openpyxl
